refactor: move files in Playground of the editor
