+++ components/Playground/NavEditor/naveditor.module.css
@@ -23,7 +23,7 @@
   border-radius: 50%;
   width: 50%;
   height: 50%;
-  background-image: -webkit-linear-gradient(120deg, var(--md-cursor), var(--md-sage));
+  background-image: -webkit-linear-gradient(120deg, var(--md-cursor), var(--md-coral) 120%);
   filter: blur(5px);
   opacity: 0;
   transition: opacity .1s ease-out;
@@ -78,6 +78,7 @@
   transition: opacity .0s ease-out;
   z-index: 1;
   opacity: 0;
+  display: none;
   border-collapse: collapse;
   display: none;
 }
@@ -110,7 +111,6 @@
   height: 100%;
   width: 100%;
   z-index: 10;
-  /* background-color: rgba(0, 0, 0, 0.603); */
 }
 
 .menu {
@@ -196,6 +196,7 @@
     width:  48px;
     position: relative;
     border-right: 1px solid var(--md-divider-light-1);
+    border-bottom: 0;
   }
 
   .navEditor {
@@ -212,12 +213,9 @@
     width: 100%;
   }
 
-  .tooltip::before, .tooltip::after {
-    display: block;
-  }
-
   .tooltip:hover::before, .tooltip:hover::after {
     opacity: 1;
+    display: flex;
     transition-delay: 0.3s;
   }
