+++ components/Playground/Preview/Preview.jsx
@@ -2,7 +2,7 @@ import React, { useEffect, useRef } from 'react'
 import highlightjs from 'highlight.js'
 import style from './preview.module.css'
 import { marked } from 'marked'
-import useThemes from '../../hooks/useThemes'
+import useThemes from '../../../hooks/useThemes'
 // import 'highlight.js/styles/github-dark.css'
 // import 'highlight.js/styles/atom-one-dark.css'
 // import 'highlight.js/styles/base16/onedark.css'
